Authenticate Docker with GitHub Packages
When a GitHub Personal Access Token (PAT) is provided for authentication
with the GitHub Packages container registry, automatically log in with
the application repository account and the provided token. Since the
authentication happens before every operation, it is still preferred to
instead authenticate the Docker engine manually before running commands.